cambios

+++ src/app/core/application/use-cases/auth/logout.use-case.ts
@@ -1,6 +1,6 @@
 import { inject, Injectable } from '@angular/core';
 import { Observable, tap } from 'rxjs';
-import { AuthRepository } from '../../../domain/repositories/auth.repository';
+import { AuthRepository, ILogoutResponse } from '../../../domain/repositories/auth.repository';
 import { StorageRepository } from '../../../domain/repositories/storage.repository';
 
 @Injectable({
@@ -10,12 +10,16 @@ export class LogoutUseCase {
   private authRepository = inject(AuthRepository);
   private storageRepository = inject(StorageRepository);
 
-  execute(): Observable<void> {
+  execute(): Observable<ILogoutResponse> {
     return this.authRepository.logout().pipe(
-      tap(() => {
-        this.storageRepository.removeItem('accessToken');
-        this.storageRepository.removeItem('refreshToken');
-        this.storageRepository.removeItem('user');
+      tap((response) => {
+        // Solo limpiar si el logout fue exitoso
+        if (response.success) {
+          this.storageRepository.removeItem('accessToken');
+          this.storageRepository.removeItem('refreshToken');
+          this.storageRepository.removeItem('user');
+          console.log('Logout successful:', response.message);
+        }
       })
     );
   }
